Refactored login component into registration module
