+++ UI/Web/src/app/registration/registration.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ConfirmEmailComponent } from './confirm-email/confirm-email.component';
 import { RegistrationRoutingModule } from './registration.router.module';
-import { NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
+import { NgbCollapseModule, NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
 import { ReactiveFormsModule } from '@angular/forms';
 import { SplashContainerComponent } from './splash-container/splash-container.component';
 import { RegisterComponent } from './register/register.component';
@@ -10,6 +10,7 @@ import { AddEmailToAccountMigrationModalComponent } from './add-email-to-account
 import { ConfirmMigrationEmailComponent } from './confirm-migration-email/confirm-migration-email.component';
 import { ResetPasswordComponent } from './reset-password/reset-password.component';
 import { ConfirmResetPasswordComponent } from './confirm-reset-password/confirm-reset-password.component';
+import { UserLoginComponent } from './user-login/user-login.component';
 
 
 
@@ -21,7 +22,8 @@ import { ConfirmResetPasswordComponent } from './confirm-reset-password/confirm-
     AddEmailToAccountMigrationModalComponent,
     ConfirmMigrationEmailComponent,
     ResetPasswordComponent,
-    ConfirmResetPasswordComponent
+    ConfirmResetPasswordComponent,
+    UserLoginComponent
   ],
   imports: [
     CommonModule,
